♻️ Refactor: 집안일 추가 스텝 2 로직, 스타일 분리
