+++ src/components/housework/index.ts
@@ -2,3 +2,6 @@ export { default as HeaderWithTitle } from './HeaderWithTitle/HeaderWithTitle';
 export { default as HouseWorkSheet } from './HouseWorkSheet/HouseWorkSheet';
 export { default as DueDateSheet } from './DueDateSheet/DueDateSheet';
 export { default as HouseworkForm } from './HouseworkForm/HouseworkForm';
+export { default as TaskAssignmentContent } from './TaskAssignmentContent/TaskAssignmentContent';
+export { default as ManagerSelectSheet } from './ManagerSelectSheet/ManagerSelectSheet';
+export { default as HouseWorkAddLoading } from './HouseworkAddLoading/HouseWorkAddLoading';
